# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Akmatori API
  description: |
    AI-powered AIOps platform API. Manages skills, tools, incidents, alert sources,
    and system settings for automated incident investigation and remediation.
  version: 1.0.0
  contact:
    name: Akmatori
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          format: int64
        total_pages:
          type: integer
...
    Incident:
      type: object
      properties:
        id:
          type: integer
        uuid:
          type: string
          format: uuid
        source:
          type: string
        source_id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [pending, running, diagnosed, observing, completed, failed]
        context:
          type: object
        session_id:
          type: string
        working_dir:
          type: string
        full_log:
          type: string
        response:
          type: string
        tokens_used:
          type: integer
        execution_time_ms:
          type: integer
          format: int64
        alert_count:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ContextFile:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        original_name:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /incidents:
    get:
      summary: List incidents
      operationId: listIncidents
      tags: [Incidents]
      parameters:
        - name: from
          in: query
          schema:
            type: integer
            format: int64
          description: Start time (unix seconds)
        - name: to
          in: query
          schema:
            type: integer
            format: int64
          description: End time (unix seconds)
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Page number (enables pagination)
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
          description: Items per page (default 50, max 200)
      responses:
        '200':
          description: |
            Without pagination params: returns array of incidents (max 500).
            With page/per_page: returns paginated envelope.
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Incident'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
    post:
      summary: Create incident
      operationId: createIncident
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task:
                  type: string
                context:
                  type: object
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                type: object
                properties:
                  uuid:
                    type: string
                  status:
                    type: string
                  working_dir:
                    type: string
                  message:
                    type: string
  /incidents/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get incident
      operationId: getIncident
      tags: [Incidents]
      responses:
        '200':
          description: Incident details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /context:
    get:
      summary: List context files
      operationId: listContextFiles
      tags: [Context]
      responses:
        '200':
          description: Context files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContextFile'
    post:
      summary: Upload context file
      operationId: uploadContextFile
      tags: [Context]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, filename]
              properties:
                file:
                  type: string
                  format: binary
                filename:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextFile'
...
  /context/{id}/download:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Download context file
      operationId: downloadContextFile
      tags: [Context]
      responses:
        '200':
          description: File content
          content:
            '*/*':
              schema:
                type: string
                format: binary
  /context/validate:
    post:
      summary: Validate file references
      operationId: validateReferences
      tags: [Context]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  references:
                    type: array
                    items:
                      type: string
                  found:
                    type: array
                    items:
                      type: string
                  missing:
                    type: array
                    items:
                      type: string

  # ===== Alert Sources =====
...